requirements.txt
torch
numpy
joblib
rdkit
pyyaml
addict
scikit-learn
tqdm
numba
huggingface_hub
pandas
umap-learn
seaborn